requirements.txt
discord.py
docker
